Add project structure.
Initial version of project layout and build system.
The build system is composed of a `configure` script (written in
POSIX shell script) that detect system, architecture and versions.
The goal of `configure` is just generate a common.mk file that could
be included from the subsystems (acorn, oak, nut, and so on).

Signed-off-by: Tiago de Bem Natel de Moura <tiago4orion@gmail.com>>